[MONDRIAN-1836] move adomd tests to XmlaAdomdTest
